requirements.txt
# Existing - Core dependencies
httpx>=0.25.0
python-dotenv>=1.0.0
pypdf>=4.0.0

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic-settings>=2.1.0

# Auth & Database
PyJWT>=2.8.0
supabase>=2.3.0
